Upgrade - Transition to TF 2.17.1 (Keras 3) as part of the migration to Python 3.11

+++ src/AlignAIR/Data/tokenizers/center_padded_sequence_tokenizer.py
@@ -38,8 +38,7 @@ class CenterPaddedSequenceTokenizer:
         Returns:
             np.ndarray: Encoded sequence.
         """
-        # Emit int32 token ids for compatibility with Keras Embedding and SavedModel signature
-        return np.array([self.token_dict.get(nt, self.token_dict['N']) for nt in sequence], dtype=np.int32)
+        return np.array([self.token_dict.get(nt, self.token_dict['N']) for nt in sequence])
 
     def encode_and_pad_center(self, sequence_or_sequences):
         """
@@ -60,7 +59,7 @@ class CenterPaddedSequenceTokenizer:
             padding_length = self.max_length - len(encoded)
             pad_left = padding_length // 2
             pad_right = padding_length - pad_left
-            padded = np.pad(encoded, (pad_left, pad_right), constant_values=0).astype(np.int32, copy=False)
+            padded = np.pad(encoded, (pad_left, pad_right), constant_values=0)
             return padded, pad_left
 
         # If it's a list/iterable of sequences
@@ -72,10 +71,10 @@ class CenterPaddedSequenceTokenizer:
             padding_length = self.max_length - len(encoded)
             pad_left = padding_length // 2
             pad_right = padding_length - pad_left
-            padded = np.pad(encoded, (pad_left, pad_right), constant_values=0).astype(np.int32, copy=False)
+            padded = np.pad(encoded, (pad_left, pad_right), constant_values=0)
             padded_batch.append(padded)
             paddings.append(pad_left)
 
-        return np.vstack(padded_batch).astype(np.int32, copy=False), np.array(paddings, dtype=np.int32)
+        return np.vstack(padded_batch), np.array(paddings)
